% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DawnRank-package.R
\docType{data}
\name{brcaExampleMutation}
\alias{brcaExampleMutation}
\title{brcaExampleMutation}
\format{The format is: num [1:1492, 1:533] 0 0 0 0 0 0 0 0 0 0 ...  -
attr(*, "dimnames")=List of 2 ..$ : chr [1:1492] "PML" "FGF3" "HLA-A"
"CACNA1F" ...  ..$ : chr [1:533] "TCGA-A1-A0SD" "TCGA-A1-A0SE"
"TCGA-A1-A0SH" "TCGA-A1-A0SJ" ...}
\source{
https://tcga-data.nci.nih.gov/tcga/dataAccessMatrix.htm
}
\description{
Mutation data for an example Breast Cancer Dataset
}
\details{
Mutation data for an example Breast Cancer Dataset. Rows are genes, columns
are patients
}
\examples{

data(brcaExampleMutation)
## maybe str(brcaExampleMutation) ; plot(brcaExampleMutation) ...

}
\references{
TCGA
}
\keyword{datasets}

