% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DawnRank-package.R
\docType{data}
\name{brcaExampleNormalExpression}
\alias{brcaExampleNormalExpression}
\title{brcaExampleNormalExpression}
\format{The format is: num [1:1492, 1:62] 0.24 0.122 0.56 -0.231 1.189 ...
- attr(*, "dimnames")=List of 2 ..$ : chr [1:1492] "PML" "FGF3" "HLA-A"
"CACNA1F" ...  ..$ : chr [1:62] "TCGA-A7-A0CE" "TCGA-A7-A0CH" "TCGA-A7-A0D9"
"TCGA-A7-A0DB" ...}
\source{
https://tcga-data.nci.nih.gov/tcga/dataAccessMatrix.htm
}
\description{
Normal Expression data for an example Breast Cancer Dataset
}
\details{
Expression data for an example Breast Cancer Dataset. Rows are genes,
columns are patients
}
\examples{

data(brcaExampleNormalExpression)
## maybe str(brcaExampleNormalExpression) ; plot(brcaExampleNormalExpression) ...

}
\references{
TCGA
}
\keyword{datasets}

