% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DawnRank-package.R
\docType{data}
\name{brcaExamplePathway}
\alias{brcaExamplePathway}
\title{brcaExamplePathway}
\format{The format is: num [1:1492, 1:1492] 0 0 0 0 0 0 0 0 0 0 ...  -
attr(*, "dimnames")=List of 2 ..$ : chr [1:1492] "PML" "FGF3" "HLA-A"
"CACNA1F" ...  ..$ : chr [1:1492] "PML" "FGF3" "HLA-A" "CACNA1F" ...}
\source{
http://www.genome.jp/kegg/
}
\description{
Pathway data for an example Breast Cancer Dataset
}
\details{
An adjacency matrix detailing interactions of an example pathway
}
\examples{

data(brcaExamplePathway)
## maybe str(brcaExamplePathway) ; plot(brcaExamplePathway) ...

}
\references{
KEGG
}
\keyword{datasets}

