% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dawn.R
\name{Dawn}
\alias{Dawn}
\title{Dawn}
\usage{
Dawn(dawnMatrix, expressionVector, mutationVector, damping, maxit = 100,
  epsilon = 1e-04, goldStandard = NULL, patientTag = "defaultPatient")
}
\arguments{
\item{dawnMatrix, }{the weighted adjacency}

\item{expressionVector, }{the normalized expression vector}

\item{mutationVector, }{a logical vector containing mutation information}

\item{damping, }{the damping vector}

\item{maxit, }{the maximum number of iterations to use, default 100}

\item{epsilon, }{the lower magnitude cutoff, default 0.0001}

\item{goldStandard, }{A list of common driver genes, used as a comparison.
This is optional, default=NULL}

\item{patientTag, }{an index for the patients}
}
\value{
the ranks. A list of 3 including a [[1]] output of all the ranks,
[[2]] mutated ranks, [[3]] the steps of convergence
}
\description{
The most basic version of DawnRank. This is DawnRank for one Patient. Dawn
differs from the other methods from calling DawnRank by being only for one
patient and not including the matrix information. Nonetheless it is still
quite useful for a quick hash for ranking a patient
}

