% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DawnMatrix.R
\name{DawnMatrix}
\alias{DawnMatrix}
\title{DawnMatrix}
\usage{
DawnMatrix(adjMatrix)
}
\arguments{
\item{adjMatrix, }{the adjacency matrix}
}
\value{
the transition matrix
}
\description{
DawnMatrix creates a transition matrix for DawnRank
}

