% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DawnRank-package.R
\docType{data}
\name{goldStandard}
\alias{goldStandard}
\title{goldStandard}
\format{The format is: chr [1:147] "ABL1" "ABL2" "ACSL3" "AKAP9" "AKT1"
"AKT2" "APC" "ARID1A" ...}
\source{
http://cancer.sanger.ac.uk/cancergenome/projects/census/
}
\description{
The Cancer Gene Census Gold standard for common driver genes.
}
\details{
The Cancer Gene Census Gold standard for common driver genes.This is a
vector
}
\examples{

data(goldStandard)
## maybe str(goldStandard) ; plot(goldStandard) ...

}
\references{
COSMIC
}
\keyword{datasets}

