% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DawnRank-package.R
\docType{data}
\name{brcaExampleTumorExpression}
\alias{brcaExampleTumorExpression}
\title{brcaExampleTumorExpression}
\format{The format is: num [1:1492, 1:533] 0.0578 0.3625 0.7008 0.3775
0.4605 ...  - attr(*, "dimnames")=List of 2 ..$ : chr [1:1492] "PML" "FGF3"
"HLA-A" "CACNA1F" ...  ..$ : chr [1:533] "TCGA-A1-A0SD" "TCGA-A1-A0SE"
"TCGA-A1-A0SH" "TCGA-A1-A0SJ" ...}
\source{
https://tcga-data.nci.nih.gov/tcga/dataAccessMatrix.htm
}
\description{
Tumor Expression data for an example Breast Cancer Dataset
}
\details{
Expression data for an example Breast Cancer Dataset. Rows are genes,
columns are patients
}
\examples{

data(brcaExampleTumorExpression)
## maybe str(brcaExampleTumorExpression) ; plot(brcaExampleTumorExpression) ...

}
\references{
TCGA
}
\keyword{datasets}

