% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dawnDamping.R
\name{dawnDamping}
\alias{dawnDamping}
\title{dawnDamping}
\usage{
dawnDamping(adjMatrix, mu)
}
\arguments{
\item{adjMatrix, }{the adjacency matrix}

\item{mu, }{the free parameter to fit}
}
\value{
the vector for individualized damping factors
}
\description{
dawnDamping calculates the damping factor
}

