\name{goldStandard}
\alias{goldStandard}
\docType{data}
\title{
goldStandard
}
\description{
The Cancer Gene Census (CGC) Gold standard for common driver genes. The CGC is maintained by COSMIC.
}
\usage{data(goldStandard)}
\format{
  The format is:
 chr [1:147] "ABL1" "ABL2" "ACSL3" "AKAP9" "AKT1" "AKT2" "APC" "ARID1A" ...
}
\details{
The Cancer Gene Census Gold standard for common driver genes.This is a vector
}
\source{
http://cancer.sanger.ac.uk/cancergenome/projects/census/
}
\references{
Futreal P, Coin L, Marshall M, Down T, Hubbard T, Wooster R, Rahman N, Stratton M: A census of human cancer genes.
Nature Reviews Cancer 2004, 4:177-183}
\examples{
data(goldStandard)
## maybe str(goldStandard) ; plot(goldStandard) ...
}
\keyword{datasets}
