% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patspeccutoff.R
\name{patspeccutoff}
\alias{patspeccutoff}
\title{patspeccuttoff}
\usage{
patspeccutoff(patient, ms, default = 0.95)
}
\arguments{
\item{patient, }{the patient of interest}

\item{ms, }{a dataframe containing dawn rank output (specifically, the third
element of DawnRank output)}

\item{default, }{the default cutoff in case there are not enough common
drivers to make a clear conclusion}
}
\value{
a data frame containing information regarding the genetic
information of a specific patient, and the cutoff itself
}
\description{
patspeccuttoff determines the patient specific cutoffs give
}

